Fix - bug with imports, after i separate displayCategories function into another file
